Fix scene view canvas/platform view placement.
